feat(urlCleaner): implement URL cleaning functionality and clipboard interceptor
