+++ src/inject/index.js
@@ -42,6 +42,84 @@ injectFunction(
   },
 )
 
+const PARAMS_TO_REMOVE = [
+  'spm_id_from',
+  'from_source',
+  'msource',
+  'bsource',
+  'seid',
+  'source',
+  'session_id',
+  'visit_id',
+  'sourceFrom',
+  'from_spmid',
+  'share_source',
+  'share_medium',
+  'share_plat',
+  'share_session_id',
+  'share_tag',
+  'unique_k',
+  'csource',
+  'vd_source',
+  'tab',
+  'trackid',
+  'is_story_h5',
+  'share_from',
+  'plat_id',
+  '-Arouter',
+  'launch_id',
+  'live_from',
+  'hotRank',
+  'broadcast_type',
+]
+
+function cleanUrl(url) {
+  try {
+    const urlObj = new URL(url)
+    let hasChanged = false
+
+    PARAMS_TO_REMOVE.forEach((param) => {
+      if (urlObj.searchParams.has(param)) {
+        urlObj.searchParams.delete(param)
+        hasChanged = true
+      }
+    })
+
+    if (hasChanged) {
+      return urlObj.toString()
+        .replace(/([^:])\/\//g, '$1/')
+        .replace(/%3D/gi, '=')
+        .replace(/%26/g, '&')
+    }
+  }
+  catch {
+  }
+  return url
+}
+
+function cleanText(text) {
+  if (!text)
+    return text
+  // 匹配 http/https 链接
+  return text.replace(/(https?:\/\/\S+)/g, (match) => {
+    return cleanUrl(match)
+  })
+}
+
+function setupClipboardInterceptor() {
+  const originalWriteText = navigator.clipboard.writeText.bind(navigator.clipboard)
+
+  navigator.clipboard.writeText = function (text) {
+    const shouldClean = document.documentElement.getAttribute('data-bewly-clean-url') === 'true'
+    if (shouldClean && typeof text === 'string' && text.includes('bilibili.com')) {
+      text = cleanText(text)
+    }
+    return originalWriteText(text)
+  }
+}
+
+setupClipboardInterceptor()
+
 window.___inject = true
 
 // History.prototype.pushState = history.pushState
